requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
nut2==2.1.1
paramiko==2.9.2
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
wakeonlan==2.1.0
